Refactor Plane/Volume to own submodule (`image`)
