+++ nanomesh/__init__.py
@@ -1,14 +1,13 @@
 import logging
 import sys
 
+from .image import Plane, Volume
 from .mesh import LineMesh, TetraMesh, TriangleMesh
 from .mesh2d import Mesher2D, compare_mesh_with_image
 from .mesh3d import Mesher3D
 from .mesh_container import MeshContainer
-from .plane import Plane
 from .region_markers import RegionMarker
 from .triangulate import simple_triangulate
-from .volume import Volume
 
 logging.basicConfig(format='%(message)s',
                     level=logging.INFO,
