+++ nanomesh/image/__init__.py
@@ -1,0 +1,10 @@
+from ._plane import Plane
+from ._utils import SliceViewer, show_image
+from ._volume import Volume
+
+__all__ = [
+    'Volume',
+    'Plane',
+    'show_image',
+    'SliceViewer',
+]
